GO-46 add main page add main controller + service

+++ src/main/java/com/innopolis/eventgo/restcontroller/PostController.java
@@ -51,7 +51,7 @@ public class PostController {
     @GetMapping("/post")
     public List<PostDto> getPostsByFilter(@RequestParam(required = false) Optional<String> city,
                                           @RequestParam(required = false) Optional<String> category,
-                                          @RequestParam(required = false) Optional<Long> statusId,
+                                          @RequestParam(required = false) Optional<Integer> statusId,
                                           @RequestParam(required = false) Optional<Integer> page,
                                           @RequestParam(required = false) Optional<Integer> size,
                                           @RequestParam(required = false) Optional<String> sort) throws PostNotFoundException {
